RN-190: Added support for using '*' in matrix column names array (#3415)
* RN-190: Added support for using '*' in matrix column names array
- Can be used to achieve ordering of columns even if some of them are dynamic
- Moved tests from output.test.ts -> matrix.test.ts as all tests were for the matrix output type

* RN-190: Made includeFields intialisation logic a bit cleaner
