+++ packages/report-server/src/reportBuilder/output/functions/matrix/types.ts
@@ -1,7 +1,7 @@
 import { Row } from '../../../types';
 
 export type MatrixParams = {
-  columns: { includeFields: string[] | '*'; excludeFields: string[] };
+  columns: { includeFields: string[]; excludeFields: string[] };
   rows: { rowField: string; categoryField: string | undefined };
 };
